requirements.txt
pytest==8.3.3
selenium==4.38.0
allure-pytest==2.13.5
allure-python-commons==2.13.5
python-dotenv==1.2.1
